move instances HasCitations1 NValue to the parent
